Move phase functions to utils submodule.
